+++ dubbo-samples-generic/src/main/java/org/apache/dubbo/samples/generic/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.generic;
+package org.apache.dubbo.samples.generic;
 
 import java.io.File;
 import java.lang.reflect.Method;
