+++ dubbo-samples-generic/src/main/java/org/apache/dubbo/samples/generic/GenericConsumer.java
@@ -17,20 +17,23 @@
  *
  */
 
-package com.alibaba.dubbo.samples.generic;
+package org.apache.dubbo.samples.generic;
+
+import org.apache.dubbo.samples.generic.api.IUserService;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 /**
- * GenericProvider
+ * GenericConsumer
  */
-public class GenericProvider {
+public class GenericConsumer {
 
     public static void main(String[] args) throws Exception {
-        new EmbeddedZooKeeper(2181, false).start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/generic-provider.xml"});
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/generic-consumer.xml"});
         context.start();
+        IUserService userservice = (IUserService) context.getBean("userservice");
+        IUserService.User user = userservice.get(new IUserService.Params("a=b"));
+        System.out.println(user);
         System.in.read();
     }
-
 }
