+++ dubbo-samples-generic/src/main/java/org/apache/dubbo/samples/generic/GenericProvider.java
@@ -17,8 +17,20 @@
  *
  */
 
-package com.alibaba.dubbo.samples.generic.api;
+package org.apache.dubbo.samples.generic;
 
-public interface IService<P, V> {
-    V get(P params);
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+/**
+ * GenericProvider
+ */
+public class GenericProvider {
+
+    public static void main(String[] args) throws Exception {
+        new EmbeddedZooKeeper(2181, false).start();
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/generic-provider.xml"});
+        context.start();
+        System.in.read();
+    }
+
 }
