+++ dubbo-samples-generic/src/main/resources/spring/generic-provider.xml
@@ -30,8 +30,8 @@
 
     <dubbo:protocol name="dubbo" port="20880"/>
 
-    <bean id="userserviceimpl" class="com.alibaba.dubbo.samples.generic.impl.UserServiceImpl"/>
+    <bean id="userserviceimpl" class="org.apache.dubbo.samples.generic.impl.UserServiceImpl"/>
 
-    <dubbo:service interface="com.alibaba.dubbo.samples.generic.api.IUserService" ref="userserviceimpl"/>
+    <dubbo:service interface="org.apache.dubbo.samples.generic.api.IUserService" ref="userserviceimpl"/>
 
 </beans>
